Move tensor related ops outside of base Problem class
