+++ experiments/prioritized_dqn/cartpole_v1.py
@@ -1,4 +1,3 @@
-import gym
 import torchrl.registry as registry
 import torchrl.utils as utils
 from torchrl.problems import PrioritizedDQNProblem
