+++ torchrl/agents/ppo_agent.py
@@ -46,7 +46,7 @@ class BasePPOAgent(BaseAgent):
   def act(self, obs):
     obs_tensor = self.obs_to_tensor(obs)
 
-    _, dist = self.ac_net(obs)
+    _, dist = self.ac_net(obs_tensor)
     action = dist.sample()
     return action.unsqueeze(1).cpu().numpy()
