+++ torchrl/problems/ddpg.py
@@ -11,7 +11,8 @@ class DDPGProblem(GymProblem):
     self.buffer = ReplayBuffer(self.hparams.buffer_size)
 
   def train(self, history_list: list):
-    # Populate the buffer
+    history_list = self.hist_to_tensor(history_list, device=torch.device('cpu'))
+
     batch_history = self.merge_histories(*history_list)
     transitions = list(zip(*batch_history))
     self.buffer.extend(transitions)
