+++ torchrl/problems/ppo.py
@@ -4,10 +4,7 @@ from ..utils import minibatch_generator
 
 class PPOProblem(GymProblem):
   def train(self, history_list: list):
-    history_list = [
-        tuple([item.to(self.device) for item in history])
-        for history in history_list
-    ]
+    history_list = self.hist_to_tensor(history_list, device=self.device)
 
     batch_history = self.merge_histories(*history_list)
     data = [self.agent.compute_returns(*history) for history in history_list]
