+++ torchrl/problems/prioritized_dqn.py
@@ -14,7 +14,8 @@ class PrioritizedDQNProblem(GymProblem):
                                           num_steps=hparams.beta_anneal_steps)
 
   def train(self, history_list: list):
-    # Populate the buffer
+    history_list = self.hist_to_tensor(history_list, device=torch.device('cpu'))
+
     batch_history = self.merge_histories(*history_list)
     transitions = list(zip(*batch_history))
     self.buffer.extend(transitions)
